规则相关功能基本完成，除了排序。删除一些过时的文件
