+++ app/renderer/src/main/src/pages/mitm/MITMRule/MITMRuleConfigure/MITMRuleConfigure.module.scss
@@ -10,3 +10,9 @@
         margin-right: 1px;
     }
 }
+
+.modal-subTitle {
+    display: flex;
+    justify-content: space-between;
+    flex: 1;
+}
